curriculum/weeks-04-end-coding/model-proxy/proxy-requirements.txt
anthropic
langfuse
python-dotenv
